// Frontend/src/app/api-client/model/taskItemDto.ts
/**
 * Presentation
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TaskItemDto { 
    id?: string;
    title?: string | null;
    description?: string | null;
    dueDate?: string;
    isCompleted?: boolean;
}